requirements.txt
python-dotenv
google-genai
fastapi
uvicorn
sqlalchemy
pydantic
python-jose
passlib
bcrypt
python-multipart
